install nix doom emacs  (#10)
* Nix doom emacs work in progress

* .hmModule

* set correct username on ci runners

* use current username for bootstrap

* hacks to hopefully make bootstra work w/ nix doom

basically it appers that the issues I was having were because of
contents in my home dir which were interfering with the setup (locally)

Is this also happening on remote? I'm not totally sure, but this is my
attempt to see

* parameterize username and home dir in home.nix

I think i programmed this successfully! I cant belive it!
maybe there is hope for nix and I yet

* ignore the hosts/current symlink

* comment out shen elisp, not gonna work on CI

* add nix-profile bin to path