finished unifying TODOs in a single section
still not quite sure about this; there is something good about having
TODO right beside the relevant thing. and I can see all of the TODOs by
doing a search. But for whatever reason, I've come to dislike the way
org mode handles TODOs, and dont really have a need for any of the misc
todo-specific features, so checkboxes work for now