Cleanup workstation org (#12)
* move TODOs to checklist

* cleanup etc

* put contents of home.nix dirctly in flake.nix

indrection is only harming for now

* remove debris from snap

however this might actually be needed in some way, I do do some things
here (like chmod that user) that I don't think is important, but not
sure, either way need to pay attention to this run.

* Revert "remove debris from snap"

This reverts commit f3b22572e82c7c895871b18d56bf0d2cbcc3dd13.

* Revert "put contents of home.nix dirctly in flake.nix"

This reverts commit dc4d5cd6d11eeff74dd2914a5cf3485a871cace8.

* cleanup and documentation

* Revert "Revert "remove debris from snap""

This reverts commit ceebd9e04e8225e81b04024b8ece792b2698ef57.

* Revert "Revert "merge home.nix into flake.nix""

This reverts commit aac081cb7db0eae0dfb76c1685eee872d76c25e8.

* logging tweaks

* try embedding darwin config directly in flake.nix

* blargh

* notes

* backport changes to workstation.org

* remove hardcoded username "joel"

* output nix.conf on macos

* add post-execution setup instructions

* get rid of these backticks which break build

* remove a bunch of "do not edit" warnings

was going to do somethign else, but then decided that this just isn't
necessary for now

* finished unifying TODOs in a single section

still not quite sure about this; there is something good about having
TODO right beside the relevant thing. and I can see all of the TODOs by
doing a search. But for whatever reason, I've come to dislike the way
org mode handles TODOs, and dont really have a need for any of the misc
todo-specific features, so checkboxes work for now

* build both bww and ws programs when bootstrapping

* bootstrap-worksation should sync secrets

* reword some TODOs

* try to fix this?

* just see if sanity is whole

* try this thing again

* unify style

* add dotfiles (again, better this time)

* finish setting up new dotfiles linking scheme

* multiple expterimental features

its a shame this is going to change immediately w next branhc, but id
still like to to keep them separate

* just use heredoc method of setting nix.conf

* i forgot to tangle last run =(

* fix bug

* see if this fixes the issue